checkpoint

+++ src/main/java/com/springBoot/autoEcole/repository/UserDao.java
@@ -10,7 +10,7 @@ import com.springBoot.autoEcole.model.User;
 
 @Repository
 @Transactional
-public interface UserDao extends CrudRepository<User, Integer> {
+public interface UserDao extends CrudRepository<User, Long> {
 
 	Optional<User> findByEmail(String email);
 	boolean existsByEmail(String email);
